+++ tests/Acceptance/FoodSharePointCest.php
@@ -47,7 +47,7 @@ class FoodSharePointCest
      */
     /* public function createFoodSharePoint(AcceptanceTester $I): void
     {
-        $address = 'Teststraße 1 37073 Teststadt Deutschland';
+        $address = 'Teststraße 1';
 
         $I->login($this->responsible['email']);
         $I->amOnPage($I->foodSharePointRegionListUrl($this->testBezirk['id']));
@@ -60,8 +60,8 @@ class FoodSharePointCest
 
         // Find an address in the search field
         $I->fillField('#searchinput', $address);
-        $I->waitForElementVisible('#searchinput_listbox');
-        $I->click("//*[@id='searchinput_listbox']//*[contains(text(), 'Teststraße 1')]");
+        $I->waitForElementVisible('#suggestions option');
+        $I->click("//*[contains(text(), 'Teststraße 1')]");
 
         // Codeception's click function doesn't work with this switch checkbox. We have to click it with javascript.
         $I->executeJs('document.getElementById(\'different_location\').click()');